+++ CHANGELOG.md
@@ -6,5 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [2.3.0] - 2017-12-08
+### Added
+- `timeout` argument in `wait` and `waitStream` methods: exception will thrown after timeout
+
 ### Changed
 - `babel-cli` -> `Rollup`
+- `timeout` argument -> `interval` in `wait` and `waitStream` methods